Conflicts:
	project/VS2010Express/XBMC.vcxproj.filters

+++ lib/ffmpeg/build_xbmc_win32.sh
@@ -1,13 +1,16 @@
 #!/bin/bash
 
-if [ -d .libs ]
+if [ "$1" == "clean" ]
 then
-rm -r .libs
+  if [ -d .libs ]
+  then
+    rm -r .libs
+  fi
+  make distclean
 fi
 
-if [ -f config.mak ]
-then
-make distclean
+if [ ! -d .libs ]; then
+  mkdir .libs
 fi
 
 OPTIONS="
@@ -37,8 +40,7 @@ OPTIONS="
 
 ./configure --extra-cflags="-fno-common -Iinclude-xbmc-win32/dxva2" --extra-ldflags="-L/xbmc/system/players/dvdplayer" ${OPTIONS} &&
  
-make && 
-mkdir .libs &&
+make &&
 cp lib*/*.dll .libs/ &&
 mv .libs/swscale-0.dll .libs/swscale-0.6.1.dll &&
 cp .libs/avcodec-52.dll /xbmc/system/players/dvdplayer/ &&
